Create Verilog_Code
In this code, we have defined two input pins for IR sensors (IR1 and IR2) and one output pin for an LED (LED). The count variable is used to store the number of visitors. The code uses external interrupts (INT0 and INT1) to detect the rising edges of the signals generated by the IR sensors.

The delay() function is used to provide a delay of 1 second between LED on and off cycles. The main function sets up the interrupt registers and enters an infinite loop that toggles the LED on and off.

The interrupt service routines ISR_IR1() and ISR_IR2() are called whenever an interrupt occurs on the input pins of the IR sensors. ISR_IR1() increments the count variable when a person enters, and ISR_IR2() decrements the count variable when a person leaves.

You will need to connect the IR sensors and LED to the appropriate pins on the microcontroller and adjust the delay time based on your specific requirements.